Put both users and projects in the graph.
This avoids generating O(N^2) edges for a project with N users. For
example, rails/rails has 697 contributors as of now, and so would have
ended up causing 697*696*2=970224 edges to be generated. By having
heterogeneous nodes in the graph, this is reduced to 697*2=1394
edges. It does cause one additional node to be in the graph, but it's
well worth it.