Generate a .clj file with the DB connection info in it.
This provides an answer* to the question of how to deploy a servlet
that knows where the database is. In this case, prior to building the
war file, Rake generates src/org/andcheese/wycatsnumber/db.clj that
just defines a map containing the DB connection info. This information
comes out of config/database.yml.

The upside of this is that the servlet doesn't have to do anything
fancy to find the database, which is good since asking the question
"how do I pass a deployment-specific string to a Java servlet?" leads
you into a bottomless pool of alphabet soup.

The downside of this is that, in order to build the war file for
e.g. production, you have to have the production DB credentials in
your config/database.yml. Also, you must remember to re-run 'rake
db_clj[development]' after building the war file, as the 'uberwar'
rake task does not restore the db.clj file to its former state.

* not necessarily a good one, mind you