Tweak user-walking logic in WalkRepo job.
This should result in far fewer useless WalkUser jobs that bail out
because the user is too fresh to bother with.

There is the possibility that this will result in missing some
refreshes that would otherwise have happened. For example, if an
author is nearly (but not quite) stale, then the old logic would have
enqueued a WalkUser job for that user, and if enough time passed
before the Resque workers got to it, then that author would have been
refreshed. Now, they'll be skipped over.

Since we've got cronjobs that periodically walk stale authors and
projects, though, the extra staleness isn't too bad, and the reduction
in queue sizes should be worth it.