Add GET /friends/:author and GET /foaf/:author routes.
They return the author's friends (Github collaborators) and
friends-of-friends, respectively.

Each route has another version with /:min-weight tacked onto it; like
GET /path, this extra parameter is the minimum number of commits
necessary for a contribution to count.

While it would be technically very easy to add routes allowing for
depth-N friends instead of restricting to 1 and 2, remember that we're
using BFS to traverse the graph, and that requires keeping track of
nodes you've seen already, which means you need memory that's
exponential wrt the depth of the traversal. Get a few concurrent
requests of that kind, and watch your server die a swappy death.