Hardcode a small number of friends to return.
If your friend count exceeds 2^16, postgresql can't handle that many placeholders,
and the query blows up. Trying to use an integer array for that query also blows
up, though in a very Java-soupy way. I don't see a clear way to make this work,
so this hack is to just let something happen already.