Don't look up author names dynamically.
When the graph is loaded, just stick the author names and gravatar IDs in there.
That way, we don't have any screwing around with the DB, we don't have any size
restrictions on our responses, and we don't have to worry about overloading it.

The downside is that the graph almost certainly takes more memory now. Can't
win them all.