Use Clojure's builtin queues instead of vector-based ones.
Thanks to the book "The Joy of Clojure" for pointing out that Clojure
has a persistent queue type built in; there's just not a lot of
syntactic sugar around it. Also, thanks to that same book for
mentioning that subvec maintains a reference to the orginal vector, so
as long as there's any sub(subsubsub...)vectors of V around, none of
V's elements can get GCed.